# backend/docs/swagger.yaml
basePath: /
definitions:
  handler.GenerateKeyRequest:
    properties:
      expires_at:
        type: string
      name:
        type: string
      scopes:
        items:
          type: string
        type: array
    type: object
  handler.GenerateKeyResponse:
    properties:
      created_at:
        type: string
      display_hint:
        type: string
      expires_at:
        type: string
      id:
        type: string
      key:
        type: string
    type: object
  handler.UpdateKeyRequest:
    properties:
      name:
        type: string
    type: object
  handler.errorResponse:
    properties:
      error:
        type: string
    type: object
  handler.loginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  handler.signupRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  handler.successResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
  model.APIKey:
    properties:
      created_at:
        type: string
      display_hint:
        type: string
      expires_at:
        type: string
      id:
        type: string
      last_used_at:
        type: string
      name:
        type: string
      scopes:
        items:
          type: string
        type: array
      user_id:
        type: string
    type: object
  model.APIKeyStats:
    properties:
      api_key_id:
        type: string
      error_counts:
        additionalProperties:
          type: integer
        type: object
      peak_rpd:
        type: integer
      peak_rpm:
        type: integer
      success_rate:
        type: number
      total_requests:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the Velox API server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Velox API
  version: "1.0"
paths:
  /auth/api-keys:
    delete:
      description: Permanently revoke and delete an API key.
      parameters:
      - description: API Key UUID
        in: query
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Key deleted
        "400":
          description: Missing ID
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - Bearer: []
      summary: Delete API Key
      tags:
      - APIKeys
    get:
      description: Retrieve a list of all API keys created by the current user.
      produces:
      - application/json
      responses:
        "200":
          description: List of API keys
          schema:
            items:
              $ref: '#/definitions/model.APIKey'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - Bearer: []
      summary: List API Keys
      tags:
      - APIKeys
    patch:
      consumes:
      - application/json
      description: Rename an existing API key.
      parameters:
      - description: API Key UUID
        in: query
        name: id
        required: true
        type: string
      - description: Update Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateKeyRequest'
      responses:
        "204":
          description: Key updated
        "400":
          description: Missing ID or invalid JSON
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - Bearer: []
      summary: Update API Key Name
      tags:
      - APIKeys
    post:
      consumes:
      - application/json
      description: Create a new API key with specific scopes and optional expiration.
      parameters:
      - description: API Key Generation Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.GenerateKeyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: API key created
          schema:
            $ref: '#/definitions/handler.GenerateKeyResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - Bearer: []
      summary: Generate API Key
      tags:
      - APIKeys
  /auth/api-keys/stats:
    get:
      description: Retrieve performance metrics and usage peaks for a specific API
        key.
      parameters:
      - description: API Key ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: API Key Statistics
          schema:
            $ref: '#/definitions/model.APIKeyStats'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - Bearer: []
      summary: Get API Key Stats
      tags:
      - APIKeys
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return a JWT token.
      parameters:
      - description: Login Request
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/handler.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/handler.successResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: User Login
      tags:
      - Auth
  /auth/logout:
    post:
      description: Invalidate the session (stateless JWT logout).
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful
          schema:
            $ref: '#/definitions/handler.successResponse'
      summary: User Logout
      tags:
      - Auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Register a new user with name, email, and password.
      parameters:
      - description: Signup Request
        in: body
        name: signup
        required: true
        schema:
          $ref: '#/definitions/handler.signupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created
          schema:
            $ref: '#/definitions/handler.successResponse'
        "400":
          description: Invalid request or validation error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: User Signup
      tags:
      - Auth
  /dashboard:
    get:
      description: Fetch profile and activity data for the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: Dashboard data
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get Dashboard Data
      tags:
      - Dashboard
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and then your personal token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"